+++ CMakeLists.txt
@@ -9,10 +9,22 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra")
 add_library(internals
     src/ast/element.cpp
     src/ast/span.cpp
+    src/evaluator/expression/break.cpp
+    src/evaluator/expression/call.cpp
+    src/evaluator/expression/cond.cpp
+    src/evaluator/expression/expression.cpp
+    src/evaluator/expression/func.cpp
+    src/evaluator/expression/lambda.cpp
     src/evaluator/expression/parameters.cpp
+    src/evaluator/expression/prog.cpp
+    src/evaluator/expression/program.cpp
+    src/evaluator/expression/quote.cpp
+    src/evaluator/expression/return.cpp
+    src/evaluator/expression/setq.cpp
+    src/evaluator/expression/symbol.cpp
+    src/evaluator/expression/while.cpp
     src/evaluator/error.cpp
     src/evaluator/evaluator.cpp
-    src/evaluator/program.cpp
     src/reader/scanner.cpp
     src/reader/token.cpp
     src/reader/parser.cpp
